Try to fix Floating

+++ api/db/migrate/20180824120605_create_countries.rb
@@ -3,9 +3,9 @@ class CreateCountries < ActiveRecord::Migration[5.2]
     create_table :countries do |t|
       t.string :code
       t.string :name
-      t.double :population_density
+      t.float :population_density
       t.string :currency
-      t.double :conversion_rate_eur
+      t.float :conversion_rate_eur
 
       t.timestamps
     end
